Refactor lib.rs into focused modules following single responsibility principle
Split monolithic 2,408-line lib.rs into modular architecture:

- src/error.rs: Error handling (BufVecError enum and implementations)
- src/core.rs: Core BufVec implementation with all interfaces
- src/iter.rs: Iterator implementations (BufVecIter, BufVecPairIter)
- src/lib.rs: Public API and crate documentation

Extract tests from inline mod to integration test files:
- tests/integration_tests.rs: Basic functionality (12 tests)
- tests/iterator_tests.rs: Iterator behavior (7 tests)
- tests/dictionary_tests.rs: Dictionary interface (14 tests)
- tests/stack_tests.rs: Stack interface (7 tests)
- tests/error_tests.rs: Error handling (10 tests)

Maintain full backward compatibility with unchanged public API.
All 50 tests pass with zero clippy warnings.
